修复bug

+++ src/main/java/com/ligg/mapper/user/UserFollowMapper.java
@@ -13,14 +13,14 @@ import java.util.List;
 public interface UserFollowMapper {
     // 关注
     @Insert("insert into user_follow(user_id, follow_user_id,create_time) values(#{userId}, #{id},NOW())")
-    void userFollow(Integer userId, Long id);
+    void userFollow(Long userId, Long id);
     // 取消关注
     @Delete("delete from user_follow where user_id = #{userId} and follow_user_id = #{id}")
-    void userUnFollow(Integer userId, Long id);
+    void userUnFollow(Long userId, Long id);
 
     // 查询关注信息
     @Select("select * from user_follow where user_id = #{userId} and follow_user_id = #{id}")
-    List<UserFollow> list(Integer userId, Long id);
+    List<UserFollow> list(Long userId, Long id);
     // 查询关注列表
     @Select("select follow_user_id,create_time from user_follow where follow_user_id = #{id}")
     List<UserFollow> followList(Long id);
